+++ sdk/datafactory/arm-datafactory/src/dataFactoryManagementClient.ts
@@ -19,7 +19,9 @@ class DataFactoryManagementClient extends DataFactoryManagementClientContext {
   // Operation groups
   operations: operations.Operations;
   factories: operations.Factories;
+  exposureControl: operations.ExposureControl;
   integrationRuntimes: operations.IntegrationRuntimes;
+  integrationRuntimeObjectMetadata: operations.IntegrationRuntimeObjectMetadata;
   integrationRuntimeNodes: operations.IntegrationRuntimeNodes;
   linkedServices: operations.LinkedServices;
   datasets: operations.Datasets;
@@ -27,8 +29,10 @@ class DataFactoryManagementClient extends DataFactoryManagementClientContext {
   pipelineRuns: operations.PipelineRuns;
   activityRuns: operations.ActivityRuns;
   triggers: operations.Triggers;
-  rerunTriggers: operations.RerunTriggers;
   triggerRuns: operations.TriggerRuns;
+  rerunTriggers: operations.RerunTriggers;
+  dataFlows: operations.DataFlows;
+  dataFlowDebugSession: operations.DataFlowDebugSession;
 
   /**
    * Initializes a new instance of the DataFactoryManagementClient class.
@@ -40,7 +44,9 @@ class DataFactoryManagementClient extends DataFactoryManagementClientContext {
     super(credentials, subscriptionId, options);
     this.operations = new operations.Operations(this);
     this.factories = new operations.Factories(this);
+    this.exposureControl = new operations.ExposureControl(this);
     this.integrationRuntimes = new operations.IntegrationRuntimes(this);
+    this.integrationRuntimeObjectMetadata = new operations.IntegrationRuntimeObjectMetadata(this);
     this.integrationRuntimeNodes = new operations.IntegrationRuntimeNodes(this);
     this.linkedServices = new operations.LinkedServices(this);
     this.datasets = new operations.Datasets(this);
@@ -48,8 +54,10 @@ class DataFactoryManagementClient extends DataFactoryManagementClientContext {
     this.pipelineRuns = new operations.PipelineRuns(this);
     this.activityRuns = new operations.ActivityRuns(this);
     this.triggers = new operations.Triggers(this);
-    this.rerunTriggers = new operations.RerunTriggers(this);
     this.triggerRuns = new operations.TriggerRuns(this);
+    this.rerunTriggers = new operations.RerunTriggers(this);
+    this.dataFlows = new operations.DataFlows(this);
+    this.dataFlowDebugSession = new operations.DataFlowDebugSession(this);
   }
 }