+++ sdk/storage/storage-blob/src/Clients.ts
@@ -28,7 +28,8 @@ import {
   BlobStartCopyFromURLResponse,
   BlobAbortCopyFromURLResponse,
   BlobCopyFromURLResponse,
-  BlobSetTierResponse
+  BlobSetTierResponse,
+  ContainerEncryptionScope,
 } from "./generatedModels";
 import { AbortSignalLike } from "@azure/abort-controller";
 import { BlobDownloadResponse } from "./BlobDownloadResponse";
@@ -47,18 +48,17 @@ import {
   DEFAULT_MAX_DOWNLOAD_RETRY_REQUESTS,
   URLConstants,
   DEFAULT_BLOB_DOWNLOAD_BLOCK_BYTES,
-  DEFAULT_BLOCK_BUFFER_SIZE_BYTES
+  DEFAULT_BLOCK_BUFFER_SIZE_BYTES,
 } from "./utils/constants";
 import {
   setURLParameter,
   extractConnectionStringParts,
   appendToURLPath
 } from "./utils/utils.common";
-import { readStreamToLocalFile } from "./utils/utils.node";
+import { fsStat, readStreamToLocalFile, streamToBuffer } from "./utils/utils.node";
 import { StorageSharedKeyCredential } from "./credentials/StorageSharedKeyCredential";
 import { AnonymousCredential } from "./credentials/AnonymousCredential";
 import { Batch } from "./utils/Batch";
-import { streamToBuffer } from "./utils/utils.node";
 import { createSpan } from "./utils/tracing";
 import { HttpRequestBody } from "@azure/core-http";
 import {
@@ -83,7 +83,6 @@ import {
 import { BlockBlob } from "./generated/src/operations";
 import { Range } from "./Range";
 import { generateBlockID } from "./utils/utils.common";
-import { fsStat } from "./utils/utils.node";
 import {
   BLOCK_BLOB_MAX_STAGE_BLOCK_BYTES,
   BLOCK_BLOB_MAX_UPLOAD_BLOB_BYTES,
@@ -178,7 +177,7 @@ export interface BlobBeginCopyFromURLOptions extends BlobStartCopyFromURLOptions
  * @export
  * @interface BlobBeginCopyFromURLResponse
  */
-export interface BlobBeginCopyFromURLResponse extends BlobStartCopyFromURLResponse {}
+export interface BlobBeginCopyFromURLResponse extends BlobStartCopyFromURLResponse { }
 
 /**
  * Options to configure the {@link BlobClient.download} operation.
@@ -441,6 +440,16 @@ export interface BlobSetMetadataOptions extends CommonOptions {
    * @memberof BlobSetMetadataOptions
    */
   customerProvidedKey?: CpkInfo;
+  /**
+   * Optional. Version 2019-07-07 and later.  Specifies the name of the encryption scope to use to
+   * encrypt the data provided in the request. If not specified, encryption is performed with the
+   * default account encryption scope.  For more information, see Encryption at Rest for Azure
+   * Storage Services.
+   *
+   * @type {string}
+   * @memberof BlobSetMetadataOptions
+   */
+  encryptionScope?: string;
 }
 
 /**
@@ -599,6 +608,16 @@ export interface BlobCreateSnapshotOptions extends CommonOptions {
    * @memberof BlobCreateSnapshotOptions
    */
   customerProvidedKey?: CpkInfo;
+  /**
+   * Optional. Version 2019-07-07 and later.  Specifies the name of the encryption scope to use to
+   * encrypt the data provided in the request. If not specified, encryption is performed with the
+   * default account encryption scope.  For more information, see Encryption at Rest for Azure
+   * Storage Services.
+   *
+   * @type {string}
+   * @memberof BlobCreateSnapshotOptions
+   */
+  encryptionScope?: string;
 }
 
 /**
@@ -716,6 +735,13 @@ export interface BlobSyncCopyFromURLOptions extends CommonOptions {
    * @memberof BlobSyncCopyFromURLOptions
    */
   sourceConditions?: ModifiedAccessConditions;
+  /**
+   * Specify the md5 calculated for the range of bytes that must be read from the copy source.
+   * 
+   * @type {Uint8Array}
+   * @memberof BlobSyncCopyFromURLOptions
+   */
+  sourceContentMD5?: Uint8Array;
 }
 
 /**
@@ -1420,6 +1446,7 @@ export class BlobClient extends StorageClient {
         metadata,
         modifiedAccessConditions: options.conditions,
         cpkInfo: options.customerProvidedKey,
+        encryptionScope: options.encryptionScope,
         spanOptions
       });
     } catch (e) {
@@ -1465,6 +1492,7 @@ export class BlobClient extends StorageClient {
         metadata: options.metadata,
         modifiedAccessConditions: options.conditions,
         cpkInfo: options.customerProvidedKey,
+        encryptionScope: options.encryptionScope,
         spanOptions
       });
     } catch (e) {
@@ -1639,6 +1667,7 @@ export class BlobClient extends StorageClient {
           sourceIfNoneMatch: options.sourceConditions.ifNoneMatch,
           sourceIfUnmodifiedSince: options.sourceConditions.ifUnmodifiedSince
         },
+        sourceContentMD5: options.sourceContentMD5,
         spanOptions
       });
     } catch (e) {
@@ -2050,6 +2079,16 @@ export interface AppendBlobCreateOptions extends CommonOptions {
    * @memberof AppendBlobCreateOptions
    */
   customerProvidedKey?: CpkInfo;
+  /**
+   * Optional. Version 2019-07-07 and later.  Specifies the name of the encryption scope to use to
+   * encrypt the data provided in the request. If not specified, encryption is performed with the
+   * default account encryption scope.  For more information, see Encryption at Rest for Azure
+   * Storage Services.
+   *
+   * @type {string}
+   * @memberof AppendBlobCreateOptions
+   */
+  encryptionScope?: string;
 }
 
 /**
@@ -2108,6 +2147,16 @@ export interface AppendBlobAppendBlockOptions extends CommonOptions {
    * @memberof AppendBlobAppendBlockOptions
    */
   customerProvidedKey?: CpkInfo;
+  /**
+   * Optional. Version 2019-07-07 and later.  Specifies the name of the encryption scope to use to
+   * encrypt the data provided in the request. If not specified, encryption is performed with the
+   * default account encryption scope.  For more information, see Encryption at Rest for Azure
+   * Storage Services.
+   *
+   * @type {string}
+   * @memberof AppendBlobAppendBlockOptions
+   */
+  encryptionScope?: string;
 }
 
 /**
@@ -2168,6 +2217,16 @@ export interface AppendBlobAppendBlockFromURLOptions extends CommonOptions {
    * @memberof AppendBlobAppendBlockFromURLOptions
    */
   customerProvidedKey?: CpkInfo;
+  /**
+   * Optional. Version 2019-07-07 and later.  Specifies the name of the encryption scope to use to
+   * encrypt the data provided in the request. If not specified, encryption is performed with the
+   * default account encryption scope.  For more information, see Encryption at Rest for Azure
+   * Storage Services.
+   *
+   * @type {string}
+   * @memberof AppendBlobAppendBlockFromURLOptions
+   */
+  encryptionScope?: string;
 }
 
 /**
@@ -2375,6 +2434,7 @@ export class AppendBlobClient extends BlobClient {
         metadata: options.metadata,
         modifiedAccessConditions: options.conditions,
         cpkInfo: options.customerProvidedKey,
+        encryptionScope: options.encryptionScope,
         spanOptions
       });
     } catch (e) {
@@ -2420,6 +2480,7 @@ export class AppendBlobClient extends BlobClient {
         transactionalContentMD5: options.transactionalContentMD5,
         transactionalContentCrc64: options.transactionalContentCrc64,
         cpkInfo: options.customerProvidedKey,
+        encryptionScope: options.encryptionScope,
         spanOptions
       });
     } catch (e) {
@@ -2479,6 +2540,7 @@ export class AppendBlobClient extends BlobClient {
           sourceIfUnmodifiedSince: options.sourceConditions.ifUnmodifiedSince
         },
         cpkInfo: options.customerProvidedKey,
+        encryptionScope: options.encryptionScope,
         spanOptions
       });
     } catch (e) {
@@ -2543,6 +2605,16 @@ export interface BlockBlobUploadOptions extends CommonOptions {
    * @memberof BlockBlobUploadOptions
    */
   customerProvidedKey?: CpkInfo;
+  /**
+   * Optional. Version 2019-07-07 and later.  Specifies the name of the encryption scope to use to
+   * encrypt the data provided in the request. If not specified, encryption is performed with the
+   * default account encryption scope.  For more information, see Encryption at Rest for Azure
+   * Storage Services.
+   *
+   * @type {string}
+   * @memberof BlockBlobUploadOptions
+   */
+  encryptionScope?: string;
   /**
    * Access tier.
    * More Details - https://docs.microsoft.com/en-us/azure/storage/blobs/storage-blob-storage-tiers
@@ -2611,6 +2683,16 @@ export interface BlockBlobStageBlockOptions extends CommonOptions {
    * @memberof BlockBlobStageBlockOptions
    */
   customerProvidedKey?: CpkInfo;
+  /**
+   * Optional. Version 2019-07-07 and later.  Specifies the name of the encryption scope to use to
+   * encrypt the data provided in the request. If not specified, encryption is performed with the
+   * default account encryption scope.  For more information, see Encryption at Rest for Azure
+   * Storage Services.
+   *
+   * @type {string}
+   * @memberof BlockBlobStageBlockOptions
+   */
+  encryptionScope?: string;
 }
 
 /**
@@ -2672,6 +2754,16 @@ export interface BlockBlobStageBlockFromURLOptions extends CommonOptions {
    * @memberof BlockBlobStageBlockFromURLOptions
    */
   customerProvidedKey?: CpkInfo;
+  /**
+   * Optional. Version 2019-07-07 and later.  Specifies the name of the encryption scope to use to
+   * encrypt the data provided in the request. If not specified, encryption is performed with the
+   * default account encryption scope.  For more information, see Encryption at Rest for Azure
+   * Storage Services.
+   *
+   * @type {string}
+   * @memberof BlockBlobStageBlockFromURLOptions
+   */
+  encryptionScope?: string;
 }
 
 /**
@@ -2717,6 +2809,16 @@ export interface BlockBlobCommitBlockListOptions extends CommonOptions {
    * @memberof BlockBlobCommitBlockListOptions
    */
   customerProvidedKey?: CpkInfo;
+  /**
+   * Optional. Version 2019-07-07 and later.  Specifies the name of the encryption scope to use to
+   * encrypt the data provided in the request. If not specified, encryption is performed with the
+   * default account encryption scope.  For more information, see Encryption at Rest for Azure
+   * Storage Services.
+   *
+   * @type {string}
+   * @memberof BlockBlobCommitBlockListOptions
+   */
+  encryptionScope?: string;
   /**
    * Access tier.
    * More Details - https://docs.microsoft.com/en-us/azure/storage/blobs/storage-blob-storage-tiers
@@ -2799,6 +2901,17 @@ export interface BlockBlobUploadStreamOptions extends CommonOptions {
    * @memberof BlockBlobUploadStreamOptions
    */
   onProgress?: (progress: TransferProgressEvent) => void;
+
+  /**
+   * Optional. Version 2019-07-07 and later.  Specifies the name of the encryption scope to use to
+   * encrypt the data provided in the request. If not specified, encryption is performed with the
+   * default account encryption scope.  For more information, see Encryption at Rest for Azure
+   * Storage Services.
+   *
+   * @type {string}
+   * @memberof BlockBlobUploadStreamOptions
+   */
+  encryptionScope?: string;
 }
 /**
  * Option interface for {@link BlockBlobClient.uploadFile} and {@link BlockBlobClient.uploadSeekableStream}.
@@ -2874,6 +2987,17 @@ export interface BlockBlobParallelUploadOptions extends CommonOptions {
    * @memberof BlockBlobParallelUploadOptions
    */
   concurrency?: number;
+
+  /**
+   * Optional. Version 2019-07-07 and later.  Specifies the name of the encryption scope to use to
+   * encrypt the data provided in the request. If not specified, encryption is performed with the
+   * default account encryption scope.  For more information, see Encryption at Rest for Azure
+   * Storage Services.
+   *
+   * @type {string}
+   * @memberof BlockBlobParallelUploadOptions
+   */
+  encryptionScope?: string;
 }
 
 /**
@@ -3121,6 +3245,7 @@ export class BlockBlobClient extends BlobClient {
         modifiedAccessConditions: options.conditions,
         onUploadProgress: options.onProgress,
         cpkInfo: options.customerProvidedKey,
+        encryptionScope: options.encryptionScope,
         tier: toAccessTier(options.tier),
         spanOptions
       });
@@ -3163,6 +3288,7 @@ export class BlockBlobClient extends BlobClient {
         transactionalContentMD5: options.transactionalContentMD5,
         transactionalContentCrc64: options.transactionalContentCrc64,
         cpkInfo: options.customerProvidedKey,
+        encryptionScope: options.encryptionScope,
         spanOptions
       });
     } catch (e) {
@@ -3218,6 +3344,7 @@ export class BlockBlobClient extends BlobClient {
         sourceContentCrc64: options.sourceContentCrc64,
         sourceRange: offset === 0 && !count ? undefined : rangeToString({ offset, count }),
         cpkInfo: options.customerProvidedKey,
+        encryptionScope: options.encryptionScope,
         spanOptions
       });
     } catch (e) {
@@ -3264,6 +3391,7 @@ export class BlockBlobClient extends BlobClient {
           metadata: options.metadata,
           modifiedAccessConditions: options.conditions,
           cpkInfo: options.customerProvidedKey,
+          encryptionScope: options.encryptionScope,
           tier: toAccessTier(options.tier),
           spanOptions
         }
@@ -3447,7 +3575,7 @@ export class BlockBlobClient extends BlobClient {
       if (numBlocks > BLOCK_BLOB_MAX_BLOCKS) {
         throw new RangeError(
           `The buffer's size is too big or the BlockSize is too small;` +
-            `the number of blocks must be <= ${BLOCK_BLOB_MAX_BLOCKS}`
+          `the number of blocks must be <= ${BLOCK_BLOB_MAX_BLOCKS}`
         );
       }
 
@@ -3467,6 +3595,7 @@ export class BlockBlobClient extends BlobClient {
             await this.stageBlock(blockID, blobFactory(start, contentLength), contentLength, {
               abortSignal: options.abortSignal,
               conditions: options.conditions,
+              encryptionScope: options.encryptionScope,
               tracingOptions: { ...options!.tracingOptions, spanOptions }
             });
             // Update progress after block is successfully uploaded to server, in case of block trying
@@ -3591,6 +3720,7 @@ export class BlockBlobClient extends BlobClient {
 
           await this.stageBlock(blockID, buffer, buffer.length, {
             conditions: options.conditions,
+            encryptionScope: options.encryptionScope,
             tracingOptions: { ...options!.tracingOptions, spanOptions }
           });
 
@@ -3703,7 +3833,7 @@ export class BlockBlobClient extends BlobClient {
       if (numBlocks > BLOCK_BLOB_MAX_BLOCKS) {
         throw new RangeError(
           `The buffer's size is too big or the BlockSize is too small;` +
-            `the number of blocks must be <= ${BLOCK_BLOB_MAX_BLOCKS}`
+          `the number of blocks must be <= ${BLOCK_BLOB_MAX_BLOCKS}`
         );
       }
 
@@ -3727,6 +3857,7 @@ export class BlockBlobClient extends BlobClient {
               {
                 abortSignal: options.abortSignal,
                 conditions: options.conditions,
+                encryptionScope: options.encryptionScope,
                 tracingOptions: { ...options!.tracingOptions, spanOptions }
               }
             );
@@ -3808,6 +3939,16 @@ export interface PageBlobCreateOptions extends CommonOptions {
    */
   customerProvidedKey?: CpkInfo;
   /**
+   * Optional. Version 2019-07-07 and later.  Specifies the name of the encryption scope to use to
+   * encrypt the data provided in the request. If not specified, encryption is performed with the
+   * default account encryption scope.  For more information, see Encryption at Rest for Azure
+   * Storage Services.
+   *
+   * @type {string}
+   * @memberof PageBlobCreateOptions
+   */
+  encryptionScope?: string;
+  /**
    * Access tier.
    * More Details - https://docs.microsoft.com/en-us/azure/storage/blobs/storage-blob-storage-tiers
    *
@@ -3873,6 +4014,16 @@ export interface PageBlobUploadPagesOptions extends CommonOptions {
    * @memberof PageBlobUploadPagesOptions
    */
   customerProvidedKey?: CpkInfo;
+  /**
+   * Optional. Version 2019-07-07 and later.  Specifies the name of the encryption scope to use to
+   * encrypt the data provided in the request. If not specified, encryption is performed with the
+   * default account encryption scope.  For more information, see Encryption at Rest for Azure
+   * Storage Services.
+   *
+   * @type {string}
+   * @memberof PageBlobUploadPagesOptions
+   */
+  encryptionScope?: string;
 }
 
 /**
@@ -3904,6 +4055,16 @@ export interface PageBlobClearPagesOptions extends CommonOptions {
    * @memberof PageBlobClearPagesOptions
    */
   customerProvidedKey?: CpkInfo;
+  /**
+   * Optional. Version 2019-07-07 and later.  Specifies the name of the encryption scope to use to
+   * encrypt the data provided in the request. If not specified, encryption is performed with the
+   * default account encryption scope.  For more information, see Encryption at Rest for Azure
+   * Storage Services.
+   *
+   * @type {string}
+   * @memberof PageBlobClearPagesOptions
+   */
+  encryptionScope?: string;
 }
 
 /**
@@ -3983,6 +4144,16 @@ export interface PageBlobResizeOptions extends CommonOptions {
    * @memberof PageBlobResizeOptions
    */
   conditions?: BlobRequestConditions;
+  /**
+   * Optional. Version 2019-07-07 and later.  Specifies the name of the encryption scope to use to
+   * encrypt the data provided in the request. If not specified, encryption is performed with the
+   * default account encryption scope.  For more information, see Encryption at Rest for Azure
+   * Storage Services.
+   *
+   * @type {string}
+   * @memberof PageBlobResizeOptions
+   */
+  encryptionScope?: string;
 }
 
 /**
@@ -4091,6 +4262,16 @@ export interface PageBlobUploadPagesFromURLOptions extends CommonOptions {
    * @memberof PageBlobUploadPagesFromURLOptions
    */
   customerProvidedKey?: CpkInfo;
+  /**
+   * Optional. Version 2019-07-07 and later.  Specifies the name of the encryption scope to use to
+   * encrypt the data provided in the request. If not specified, encryption is performed with the
+   * default account encryption scope.  For more information, see Encryption at Rest for Azure
+   * Storage Services.
+   *
+   * @type {string}
+   * @memberof PageBlobUploadPagesFromURLOptions
+   */
+  encryptionScope?: string;
 }
 
 /**
@@ -4295,6 +4476,7 @@ export class PageBlobClient extends BlobClient {
         metadata: options.metadata,
         modifiedAccessConditions: options.conditions,
         cpkInfo: options.customerProvidedKey,
+        encryptionScope: options.encryptionScope,
         tier: toAccessTier(options.tier),
         spanOptions
       });
@@ -4340,6 +4522,7 @@ export class PageBlobClient extends BlobClient {
         transactionalContentMD5: options.transactionalContentMD5,
         transactionalContentCrc64: options.transactionalContentCrc64,
         cpkInfo: options.customerProvidedKey,
+        encryptionScope: options.encryptionScope,
         spanOptions
       });
     } catch (e) {
@@ -4400,6 +4583,7 @@ export class PageBlobClient extends BlobClient {
             sourceIfUnmodifiedSince: options.sourceConditions.ifUnmodifiedSince
           },
           cpkInfo: options.customerProvidedKey,
+          encryptionScope: options.encryptionScope,
           spanOptions
         }
       );
@@ -4439,6 +4623,7 @@ export class PageBlobClient extends BlobClient {
         range: rangeToString({ offset, count }),
         sequenceNumberAccessConditions: options.conditions,
         cpkInfo: options.customerProvidedKey,
+        encryptionScope: options.encryptionScope,
         spanOptions
       });
     } catch (e) {
@@ -4515,6 +4700,7 @@ export class PageBlobClient extends BlobClient {
       "PageBlobClient-getPageRangesDiff",
       options.tracingOptions
     );
+
     try {
       return await this.pageBlobContext
         .getPageRangesDiff({
@@ -4522,6 +4708,51 @@ export class PageBlobClient extends BlobClient {
           leaseAccessConditions: options.conditions,
           modifiedAccessConditions: options.conditions,
           prevsnapshot: prevSnapshot,
+          range: rangeToString({ offset, count }),
+          spanOptions
+        })
+        .then(rangeResponseFromModel);
+    } catch (e) {
+      span.setStatus({
+        code: CanonicalCode.UNKNOWN,
+        message: e.message
+      });
+      throw e;
+    } finally {
+      span.end();
+    }
+  }
+
+  /**
+  * Gets the collection of page ranges that differ between a specified snapshot and this page blob for managed disks.
+  * @see https://docs.microsoft.com/rest/api/storageservices/get-page-ranges
+  *
+  * @param {number} offset Starting byte position of the page blob
+  * @param {number} count Number of bytes to get ranges diff.
+  * @param {string} prevSnapshotUrl URL of snapshot to retrive the difference.
+  * @param {PageBlobGetPageRangesDiffOptions} [options] Options to the Page Blob Get Page Ranges Diff operation.
+  * @returns {Promise<PageBlobGetPageRangesDiffResponse>} Response data for the Page Blob Get Page Range Diff operation.
+  * @memberof PageBlobClient
+  */
+  public async getPageRangesDiffForManagedDisks(
+    offset: number,
+    count: number,
+    prevSnapshotUrl: string,
+    options: PageBlobGetPageRangesDiffOptions = {}
+  ): Promise<PageBlobGetPageRangesDiffResponse> {
+    options.conditions = options.conditions || {};
+    const { span, spanOptions } = createSpan(
+      "PageBlobClient-GetPageRangesDiffForManagedDisks",
+      options.tracingOptions
+    );
+
+    try {
+      return await this.pageBlobContext
+        .getPageRangesDiff({
+          abortSignal: options.abortSignal,
+          leaseAccessConditions: options.conditions,
+          modifiedAccessConditions: options.conditions,
+          prevSnapshotUrl,
           range: rangeToString({ offset, count }),
           spanOptions
         })
@@ -4557,6 +4788,7 @@ export class PageBlobClient extends BlobClient {
         abortSignal: options.abortSignal,
         leaseAccessConditions: options.conditions,
         modifiedAccessConditions: options.conditions,
+        encryptionScope: options.encryptionScope,
         spanOptions
       });
     } catch (e) {
@@ -5007,6 +5239,13 @@ export interface ContainerCreateOptions extends CommonOptions {
    * @memberof ContainerCreateOptions
    */
   access?: PublicAccessType;
+  /**
+   * Container encryption scope info.
+   * 
+   * @type {ContainerEncryptionScope}
+   * @memberof ContainerCreateOptions
+   */
+  containerEncryptionScope?: ContainerEncryptionScope;
 }
 
 /**
@@ -5162,24 +5401,24 @@ export interface SignedIdentifier {
 export declare type ContainerGetAccessPolicyResponse = {
   signedIdentifiers: SignedIdentifierModel[];
 } & ContainerGetAccessPolicyHeaders & {
+  /**
+   * The underlying HTTP response.
+   */
+  _response: HttpResponse & {
     /**
-     * The underlying HTTP response.
+     * The parsed HTTP response headers.
      */
-    _response: HttpResponse & {
-      /**
-       * The parsed HTTP response headers.
-       */
-      parsedHeaders: ContainerGetAccessPolicyHeaders;
-      /**
-       * The response body as text (string format)
-       */
-      bodyAsText: string;
-      /**
-       * The response body as parsed JSON or XML
-       */
-      parsedBody: SignedIdentifierModel[];
-    };
+    parsedHeaders: ContainerGetAccessPolicyHeaders;
+    /**
+     * The response body as text (string format)
+     */
+    bodyAsText: string;
+    /**
+     * The response body as parsed JSON or XML
+     */
+    parsedBody: SignedIdentifierModel[];
   };
+};
 
 /**
  * Options to configure {@link ContainerClient.setAccessPolicy} operation.