+++ sdk/storage/storage-file-share/src/utils/utils.node.ts
@@ -3,7 +3,6 @@
 
 import * as fs from "fs";
 import * as util from "util";
-import { isNode } from "@azure/core-http";
 
 /**
  * Reads a readable stream into buffer. Fill the buffer from offset to end.
@@ -99,4 +98,4 @@ export async function readStreamToLocalFile(
  *
  * Promisified version of fs.stat().
  */
-export const fsStat = util.promisify(isNode ? fs.stat : function stat() {});
+export const fsStat = util.promisify(fs.stat);