+++ sdk/storage/storage-queue/src/utils/constants.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 
 export const SDK_VERSION: string = "12.0.3";
-export const SERVICE_VERSION: string = "2019-02-02";
+export const SERVICE_VERSION: string = "2019-07-07";
 
 /**
  * The OAuth scope to use with Azure Storage.